requirements.txt
Django==1.5
django-admin-bootstrapped==1.6.9
django-bootstrap3==5.1.1
html5lib==0.9999999
Pillow==3.0.0
pisa==3.0.33
reportlab==2.6
django-mptt==0.7.4
django-media-tree	0.8.1
mock 1.3.0
